Share DialogueFlags
