+++ CSSE/CSDialogue.h
@@ -3,20 +3,14 @@
 #include "CSDefines.h"
 
 #include "CSBaseObject.h"
+
 #include "NIIteratedList.h"
+#include "TES3DialogueFlags.h"
 
 namespace se::cs {
-	enum class DialogueType : unsigned char {
-		Topic,
-		Voice,
-		Greeting,
-		Persuasion,
-		Journal
-	};
-
 	struct Dialogue : BaseObject {
 		const char* id; // 0x10
-		DialogueType type; // 0x14
+		TES3::DialogueType type; // 0x14
 		NI::IteratedList<DialogueInfo*> infos; // 0x18
 
 		bool search(std::string_view needle, const SearchSettings& settings, std::regex* regex) const;
