+++ MWSE/TES3DialogueConditional.h
@@ -1,125 +1,9 @@
 #pragma once
 
 #include "TES3Defines.h"
+#include "TES3DialogueFlags.h"
 
 namespace TES3 {
-	enum class DialogueConditionalType : unsigned char {
-		NoCondition,
-		Function,
-		GlobalVar,
-		LocalVar,
-		JournalIndex,
-		ItemCount,
-		DeadActor,
-		NotID,
-		NotFaction,
-		NotClass,
-		NotRace,
-		NotCell,
-		NotLocal,
-	};
-
-	enum class DialogueConditionalConstantType : char {
-		None = '\0',
-		Class = 'C',
-		DeadActor = 'D',
-		Faction = 'F',
-		Item = 'I',
-		Journal = 'J',
-		Cell = 'L',
-		Race = 'R',
-		NotID = 'X',
-		VariableFloat = 'f',
-		VariableInt = 'i',
-		VariableShort = 's',
-	};
-
-	enum class DialogueConditionalComparator : unsigned char {
-		Equal = 0,
-		NotEqual,
-		Greater,
-		GreaterEqual,
-		Less,
-		LessEqual,
-	};
-
-	enum class DialogueConditionalFunction : int {
-		ReactionLow = 0,
-		ReactionHigh,
-		RankRequirement,
-		Reputation,
-		HealthPercent,
-		PCReputation,
-		PCLevel,
-		PCHealthPercent,
-		PCMagicka,
-		PCFatigue,
-		PCStrength,
-		PCBlock,
-		PCArmorer,
-		PCMediumArmor,
-		PCHeavyArmor,
-		PCBluntWeapon,
-		PCLongBlade,
-		PCAxe,
-		PCSpear,
-		PCAthletics,
-		PCEnchant,
-		PCDestruction,
-		PCAlteration,
-		PCIllusion,
-		PCConjuration,
-		PCMysticism,
-		PCRestoration,
-		PCAlchemy,
-		PCUnarmored,
-		PCSecurity,
-		PCSneak,
-		PCAcrobatics,
-		PCLightArmor,
-		PCShortBlade,
-		PCMarksman,
-		PCMercantile,
-		PCSpeechcraft,
-		PCHandToHand,
-		PCSex,
-		PCExpelled,
-		PCCommonDisease,
-		PCBlightDisease,
-		PCClothingModifier,
-		PCCrimeLevel,
-		SameSex,
-		SameRace,
-		SameFaction,
-		FactionRankDifference,
-		Detected,
-		Alarmed,
-		Choice,
-		PCIntelligence,
-		PCWillpower,
-		PCAgility,
-		PCSpeed,
-		PCEndurance,
-		PCPersonality,
-		PCLuck,
-		PCCorprus,
-		Weather,
-		PCVampire,
-		Level,
-		Attacked,
-		TalkedToPC,
-		PCHealth,
-		CreatureTarget,
-		FriendHit,
-		Fight,
-		Hello,
-		Alarm,
-		Flee,
-		ShouldAttack,
-		Werewolf,
-		PCWerewolfKills,
-	};
-
 	struct DialogueConditional {
 		union {
 			BaseObject* object;
