+++ SharedSE/SharedSE.vcxitems.filters
@@ -369,6 +369,9 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3GMST.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3DialogueFlags.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
